nbaTrends-jobs/requirements-win.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
idna==2.6
py4j==0.10.6
pyspark==2.3.0
python-crontab==2.2.8
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
